# Conflicts:
#	integration-tests/src/test/java/cc/blynk/integration/web/reporting/ReportingStatsTest.java

+++ server/core/src/main/java/cc/blynk/server/core/model/widgets/ui/tiles/group/SwitchWith3LabelsGroupTemplate.java
@@ -17,7 +17,7 @@ public final class SwitchWith3LabelsGroupTemplate extends BaseGroupTemplate {
     public final GroupLabel[] groupLabels;
 
     @JsonCreator
-    public SwitchWith3LabelsGroupTemplate(@JsonProperty("id") int id,
+    public SwitchWith3LabelsGroupTemplate(@JsonProperty("id") long id,
                                           @JsonProperty("widgets") Widget[] widgets,
                                           @JsonProperty("name") String name,
                                           @JsonProperty("icon") String icon,
